+++ project/backend/demo/src/main/java/com/example/demo/controller/UserController.java
@@ -1,3 +1,5 @@
+package com.example.demo.controller; 
+
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
@@ -8,7 +10,7 @@ import java.util.Map;
 @RestController
 public class UserController {
 
-    @GetMapping("api/me")
+    @GetMapping("/api/me")
     public Map<String, Object> me(@AuthenticationPrincipal OidcUser oidcUser) {
         if (oidcUser == null) {
             return Map.of("authenticated", false);
