+++ project/backend/demo/src/main/resources/application.yml
@@ -7,7 +7,9 @@ spring:
       client:
         registration:
           google:
+            # 
             client-id: ${GOOGLE_CLIENT_ID}
+            # 
             client-secret: ${GOOGLE_CLIENT_SECRET}
             scope: openid, email, profile
             redirect-uri: "{baseUrl}/login/oauth2/code/{registrationId}"
