--- project/src/components/Login.tsx
@@ -1,190 +1,0 @@
-import React, { useState } from 'react';
-import { useNavigate, Link } from 'react-router-dom';
-import { Mail, Lock, AlertCircle, LogIn, Apple, Facebook } from 'lucide-react';
-import { useAuth } from '../contexts/AuthContext';
-import { signInWithEmailAndPassword } from 'firebase/auth';
-import { auth } from '../lib/firebase';
-
-const Login: React.FC = () => {
-  const [email, setEmail] = useState('');
-  const [password, setPassword] = useState('');
-  const [error, setError] = useState<string | null>(null);
-  const [isLoading, setIsLoading] = useState(false);
-  const [isFacebookLoading, setIsFacebookLoading] = useState(false);
-  const navigate = useNavigate();
-  const { signInWithFacebook } = useAuth();
-
-  const handleSubmit = async (e: React.FormEvent) => {
-    e.preventDefault();
-    setIsLoading(true);
-    setError(null);
-
-    try {
-      await signInWithEmailAndPassword(auth, email, password);
-      navigate('/dashboard');
-    } catch (err) {
-      setError('Invalid email or password. Please try again.');
-      console.error('Authentication error:', err);
-    } finally {
-      setIsLoading(false);
-    }
-  };
-
-  const handleFacebookLogin = async () => {
-    setIsFacebookLoading(true);
-    setError(null);
-    
-    try {
-      const result = await signInWithFacebook();
-      
-      if (result.error) {
-        setError('Facebook authentication failed. Please try again.');
-        console.error('Facebook auth error:', result.error);
-      } else if (result.user) {
-        navigate('/dashboard');
-      }
-    } catch (err) {
-      setError('Facebook authentication failed. Please try again.');
-      console.error('Facebook auth error:', err);
-    } finally {
-      setIsFacebookLoading(false);
-    }
-  };
-
-  return (
-    <div className="min-h-screen flex flex-col items-center justify-center bg-gray-50 py-12 px-4 sm:px-6 lg:px-8">
-      <div className="max-w-md w-full">
-        <div className="flex justify-center mb-8">
-          <img src="/gyb-logo.svg" alt="GYB Logo" className="h-12" />
-        </div>
-
-        <div className="bg-white shadow-md rounded-lg p-8">
-          <h2 className="text-2xl font-bold text-center text-gray-900 mb-8">
-            Welcome back!
-          </h2>
-
-          <button className="w-full bg-black text-white rounded-full py-3 px-4 font-semibold flex items-center justify-center mb-4">
-            <Apple size={24} className="mr-2" />
-            Continue with Apple
-          </button>
-          
-          <button 
-            onClick={handleFacebookLogin}
-            disabled={isFacebookLoading}
-            className="w-full bg-[#1877F2] hover:bg-[#166FE5] text-white rounded-full py-3 px-4 font-semibold flex items-center justify-center mb-4 transition-colors duration-200 disabled:opacity-50"
-          >
-            <Facebook size={24} className="mr-2" />
-            {isFacebookLoading ? 'Connecting...' : 'Continue with Facebook'}
-          </button>
-
-          <div className="flex items-center my-4">
-            <div className="flex-grow border-t border-gray-300"></div>
-            <span className="px-4 text-gray-500">or</span>
-            <div className="flex-grow border-t border-gray-300"></div>
-          </div>
-
-          <form onSubmit={handleSubmit} className="space-y-6">
-            <div>
-              <label htmlFor="email" className="block text-sm font-medium text-gray-700 mb-1">
-                Email address
-              </label>
-              <div className="relative">
-                <div className="absolute inset-y-0 left-0 pl-3 flex items-center pointer-events-none">
-                  <Mail className="h-5 w-5 text-gray-400" />
-                </div>
-                <input
-                  id="email"
-                  name="email"
-                  type="email"
-                  autoComplete="email"
-                  value={email}
-                  onChange={(e) => setEmail(e.target.value)}
-                  className="appearance-none block w-full px-3 py-2 pl-10 border border-gray-300 rounded-md shadow-sm placeholder-gray-400 focus:outline-none focus:ring-navy-blue focus:border-navy-blue sm:text-sm"
-                  placeholder="Enter your email"
-                />
-              </div>
-            </div>
-
-            <div>
-              <label htmlFor="password" className="block text-sm font-medium text-gray-700 mb-1">
-                Password
-              </label>
-              <div className="relative">
-                <div className="absolute inset-y-0 left-0 pl-3 flex items-center pointer-events-none">
-                  <Lock className="h-5 w-5 text-gray-400" />
-                </div>
-                <input
-                  id="password"
-                  name="password"
-                  type="password"
-                  autoComplete="current-password"
-                  value={password}
-                  onChange={(e) => setPassword(e.target.value)}
-                  className="appearance-none block w-full px-3 py-2 pl-10 border border-gray-300 rounded-md shadow-sm placeholder-gray-400 focus:outline-none focus:ring-navy-blue focus:border-navy-blue sm:text-sm"
-                  placeholder="Enter your password"
-                />
-              </div>
-            </div>
-
-            <div className="flex items-center justify-between">
-              <div className="flex items-center">
-                <input
-                  id="remember-me"
-                  name="remember-me"
-                  type="checkbox"
-                  className="h-4 w-4 text-navy-blue focus:ring-navy-blue border-gray-300 rounded"
-                />
-                <label htmlFor="remember-me" className="ml-2 block text-sm text-gray-700">
-                  Remember me
-                </label>
-              </div>
-              <Link
-                to="/forgot-password"
-                className="text-sm font-medium text-navy-blue hover:text-navy-blue/80"
-              >
-                Forgot password?
-              </Link>
-            </div>
-
-            {error && (
-              <div className="bg-red-50 border border-red-200 text-red-600 px-4 py-3 rounded-md flex items-center">
-                <AlertCircle className="h-5 w-5 mr-2" />
-                <span className="text-sm">{error}</span>
-              </div>
-            )}
-
-            <button
-              type="submit"
-              disabled={isLoading}
-              className="w-full flex justify-center items-center px-4 py-2 border border-transparent text-sm font-medium rounded-md text-white bg-navy-blue hover:bg-opacity-90 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-navy-blue disabled:opacity-50 disabled:cursor-not-allowed"
-            >
-              {isLoading ? (
-                <span className="flex items-center">
-                  <svg className="animate-spin -ml-1 mr-3 h-5 w-5 text-white" xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24">
-                    <circle className="opacity-25" cx="12" cy="12" r="10" stroke="currentColor" strokeWidth="4"></circle>
-                    <path className="opacity-75" fill="currentColor" d="M4 12a8 8 0 018-8V0C5.373 0 0 5.373 0 12h4zm2 5.291A7.962 7.962 0 014 12H0c0 3.042 1.135 5.824 3 7.938l3-2.647z"></path>
-                  </svg>
-                  Signing in...
-                </span>
-              ) : (
-                <span className="flex items-center">
-                  <LogIn className="mr-2" size={20} />
-                  Sign in
-                </span>
-              )}
-            </button>
-
-            <div className="text-center mt-4">
-              <span className="text-gray-600">Don't have an account?</span>
-              <Link to="/signin" className="ml-2 text-navy-blue hover:text-navy-blue/80 font-medium">
-                Sign up
-              </Link>
-            </div>
-          </form>
-        </div>
-      </div>
-    </div>
-  );
-};
-
-export default Login;
